BPSO

+++ src/org/gandhim/pso/Velocity.java
@@ -6,18 +6,18 @@ package org.gandhim.pso;
 
 public class Velocity {
 	// store the Velocity in an array to accommodate multi-dimensional problem space
-	private int[] vel;
+	private double[] vel;
 
-	public Velocity(int[] vel) {
+	public Velocity(double[] vel) {
 		super();
 		this.vel = vel;
 	}
 
-	public int[] getPos() {
+	public double[] getPos() {
 		return vel;
 	}
 
-	public void setPos(int[] vel) {
+	public void setPos(double[] vel) {
 		this.vel = vel;
 	}
